# internal/api/api.go
// Package api is the HTTP layer: routing, the response envelope, and
// translating service errors to status codes. Handlers hold no SQL and no
// filesystem access — everything goes through the service layer.
package api

import (
	_ "embed"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strings"

	"github.com/jclement/quire/internal/auth"
	"github.com/jclement/quire/internal/service"
	"github.com/jclement/quire/internal/share"
	"github.com/jclement/quire/internal/vault"
)

//go:embed openapi.yaml
var openAPISpec []byte

// Server carries the API's dependencies.
type Server struct {
	Service *service.Service
	Events  *Broadcaster
	Shares  *share.Manager
	// Auth backs the credential-management routes. Nil in tests that only
	// exercise document handlers, so those routes register only when set.
	Auth    *auth.Store
	Version string
	// Git reports whether the vault is git-backed, for the health payload.
	Git bool
	// Email, when set, answers Settings' email status and sends the test
	// digest; nil means SMTP is not configured.
	Email *EmailHooks
	// UpdateCheck reports whether a newer release exists; nil means the
	// check is disabled and health honestly says false.
	UpdateCheck func() bool
}

// Routes registers all /api/v1 handlers onto mux.
func (s *Server) Routes(mux *http.ServeMux) {
	mux.HandleFunc("GET /api/v1/health", s.handleHealth)
	mux.HandleFunc("GET /api/openapi.yaml", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/yaml")
		_, _ = w.Write(openAPISpec)
	})

	mux.HandleFunc("GET /api/v1/documents", s.handleListDocuments)
	mux.HandleFunc("POST /api/v1/documents", s.handleCreateDocument)
	mux.HandleFunc("GET /api/v1/documents/{path...}", s.handleGetDocument)
	mux.HandleFunc("PUT /api/v1/documents/{path...}", s.handleUpdateDocument)
	mux.HandleFunc("DELETE /api/v1/documents/{path...}", s.handleDeleteDocument)

	mux.HandleFunc("POST /api/v1/rename", s.handleRenameDocument)
	mux.HandleFunc("PATCH /api/v1/documents/{path...}", s.handleSetFrontmatter)
	mux.HandleFunc("POST /api/v1/link", s.handleLink)
	mux.HandleFunc("POST /api/v1/aliases", s.handleAddAlias)
	mux.HandleFunc("GET /api/v1/decisions", s.handleListDecisions)
	mux.HandleFunc("GET /api/v1/work-items", s.handleGetWorkItems)
	mux.HandleFunc("PUT /api/v1/work-items", s.handleSetWorkItems)

	mux.HandleFunc("GET /api/v1/search", s.handleSearch)
	mux.HandleFunc("GET /api/v1/related", s.handleRelated)
	mux.HandleFunc("GET /api/v1/semantic/status", s.handleSemanticStatus)
	mux.HandleFunc("GET /api/v1/timezone", s.handleGetTimezone)
	mux.HandleFunc("PUT /api/v1/timezone", s.handleSetTimezone)
	mux.HandleFunc("GET /api/v1/email", s.handleEmailStatus)
	mux.HandleFunc("POST /api/v1/email/test", s.handleEmailTest)
	mux.HandleFunc("GET /api/v1/tags", s.handleListTags)
	mux.HandleFunc("GET /api/v1/unwritten", s.handleListUnwritten)
	mux.HandleFunc("GET /api/v1/areas", s.handleListAreas)
	mux.HandleFunc("PUT /api/v1/areas", s.handleSetAreas)
	mux.HandleFunc("GET /api/v1/templates", s.handleListTemplates)
	mux.HandleFunc("POST /api/v1/templates/starter", s.handleInstallStarterTemplates)

	mux.HandleFunc("GET /api/v1/tasks", s.handleListTasks)
	mux.HandleFunc("POST /api/v1/tasks", s.handleCreateTask)
	mux.HandleFunc("POST /api/v1/tasks/{id}/toggle", s.handleToggleTask)
	mux.HandleFunc("POST /api/v1/tasks/{id}/restore-recurrence", s.handleRestoreRecurrence)
	mux.HandleFunc("PATCH /api/v1/tasks/{id}", s.handleEditTask)
	mux.HandleFunc("GET /api/v1/waiting", s.handleWaitingGroups)
	mux.HandleFunc("POST /api/v1/notes/from-task", s.handleTaskToNote)

	mux.HandleFunc("GET /api/v1/daily", s.handleListDaily)
	mux.HandleFunc("GET /api/v1/daily/{date}", s.handleGetDaily)
	mux.HandleFunc("POST /api/v1/daily/{date}", s.handleEnsureDaily)
	mux.HandleFunc("GET /api/v1/weekly/{week}", s.handleWeekReview)
	mux.HandleFunc("POST /api/v1/weekly/{week}", s.handleEnsureWeekly)
	mux.HandleFunc("GET /api/v1/today", s.handleToday)
	mux.HandleFunc("GET /api/v1/calendar", s.handleCalendar)
	s.calendarRoutes(mux)

	mux.HandleFunc("GET /api/v1/agent-guidance", s.handleGetGuidance)
	mux.HandleFunc("PUT /api/v1/agent-guidance", s.handleSetGuidance)

	mux.HandleFunc("POST /api/v1/attachments", s.handleUploadAttachment)
	mux.HandleFunc("POST /api/v1/capture", s.handleCapture)
	mux.HandleFunc("POST /api/v1/capture/note", s.handleCaptureNote)
	mux.HandleFunc("POST /api/v1/drawings", s.handleCreateDrawing)
	mux.HandleFunc("PUT /api/v1/drawings/{path...}", s.handleSaveDrawing)
	mux.HandleFunc("GET /api/v1/files/{path...}", s.handleServeFile)

	mux.HandleFunc("GET /api/v1/events", s.handleEvents)

	if s.Shares != nil {
		mux.HandleFunc("GET /api/v1/shares", s.handleListShares)
		mux.HandleFunc("POST /api/v1/shares", s.handleCreateShare)
		mux.HandleFunc("DELETE /api/v1/shares/{token}", s.handleRevokeShare)
	}

	if s.Auth != nil {
		mux.HandleFunc("GET /api/v1/tokens", s.handleListTokens)
		mux.HandleFunc("POST /api/v1/tokens", s.handleCreateToken)
		mux.HandleFunc("DELETE /api/v1/tokens/{prefix}", s.handleRevokeToken)
		mux.HandleFunc("GET /api/v1/audit", s.handleListAudit)
		mux.HandleFunc("GET /api/v1/connected-apps", s.handleListConnectedApps)
		mux.HandleFunc("DELETE /api/v1/connected-apps/{id}", s.handleDisconnectApp)
	}
}

// ---- envelope ----

type errorBody struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

func writeData(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(map[string]any{"data": data}); err != nil {
		slog.Warn("encoding response", "err", err)
	}
}

func writeError(w http.ResponseWriter, status int, code, message string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(map[string]any{"error": errorBody{Code: code, Message: message}})
}

// writeServiceError maps the service layer's sentinel errors onto HTTP
// semantics; anything unrecognized is a 500 with a generic message (never
// leak internals).
func writeServiceError(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, service.ErrValidation):
		// A caller mistake, and the message is written for the caller, so
		// pass it through rather than hiding it behind a generic 400.
		writeError(w, http.StatusBadRequest, "VALIDATION_ERROR",
			strings.TrimPrefix(err.Error(), "validation: "))
	case errors.Is(err, vault.ErrNotFound):
		writeError(w, http.StatusNotFound, "NOT_FOUND", "not found")
	case errors.Is(err, vault.ErrConflict):
		writeError(w, http.StatusConflict, "CONFLICT", "the file changed on disk — reload and reapply your edit")
	default:
		slog.Error("request failed", "err", err)
		writeError(w, http.StatusInternalServerError, "INTERNAL_ERROR", "internal error")
	}
}

func decodeBody(w http.ResponseWriter, r *http.Request, dst any) bool {
	if err := json.NewDecoder(r.Body).Decode(dst); err != nil {
		writeError(w, http.StatusBadRequest, "VALIDATION_ERROR", "invalid JSON body")
		return false
	}
	return true
}