# internal/api/openapi.yaml
# Hand-maintained OpenAPI spec, served at /api/openapi.yaml. It IS the API
# documentation — keep it truthful when handlers change.
openapi: 3.1.0
info:
  title: quire API
  version: "1"
  description: |
    quire's REST surface. Markdown on disk is canonical; every response is
    derived from it. Envelope: successes are `{"data": ...}`, errors are
    `{"error": {"code", "message"}}` with codes VALIDATION_ERROR,
    UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR.

    Authentication: session-less. In auth mode "none" (loopback), requests
    are the vault owner. Otherwise send `Authorization: Bearer sk_...`
    (create tokens with `quire token create`). Scopes: read, write, tasks.
paths:
  /api/v1/health:
    get:
      summary: Liveness, version, update availability (unauthenticated)
      responses: { "200": { description: OK } }
  /api/v1/documents:
    get:
      summary: List documents
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [note, person, company, project, meeting, daily] } }
        - { name: q, in: query, description: case-insensitive title substring, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 500 } }
      responses: { "200": { description: "data: DocMeta[]" } }
    post:
      summary: Create a typed document (server picks the path from the title)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type, title]
              properties:
                type: { type: string, enum: [note, person, company, project, meeting] }
                title: { type: string }
                markdown: { type: string, description: optional body; defaults to a heading }
      responses: { "201": { description: "data: Document" } }
  /api/v1/documents/{path}:
    get:
      summary: Full document — raw markdown plus parsed structure
      description: >
        Returns markdown, frontmatter (object), links (with resolution),
        backlinks, and contained tasks. `sha256` is the concurrency token for
        writes.
      parameters: [{ name: path, in: path, required: true, schema: { type: string }, description: "vault-relative, may contain slashes" }]
      responses: { "200": { description: "data: Document" }, "404": { description: not found } }
    put:
      summary: Replace document content (compare-and-swap)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [markdown, base_sha256]
              properties:
                markdown: { type: string }
                base_sha256: { type: string, description: sha256 from the read this edit is based on }
      responses:
        "200": { description: "data: Document" }
        "409": { description: file changed on disk — re-read and reapply }
    delete:
      summary: Delete a document
      responses: { "204": { description: deleted } }
    patch:
      summary: Surgically set or remove frontmatter keys
      description: >
        Body {set: {key: value}, base_sha256?}. A null value removes the key.
        Only the touched keys' lines change — key order, comments and quoting
        elsewhere are preserved. Strings are quoted when YAML needs it (a
        wikilink starts with `[`); arrays are written inline.
      responses: { "200": { description: "data: Document" }, "409": { description: stale base_sha256 } }
  /api/v1/search:
    get:
      summary: Search (shared grammar)
      description: "Bare words are FTS (last word prefix-matched); `type:x` and `tag:x` filter."
      parameters: [{ name: q, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: "data: SearchResult[] (snippet uses <mark>)" } }
  /api/v1/tasks:
    get:
      summary: Task views
      description: >
        inbox is open, undated, unfiled and not #someday; someday is open
        #someday tasks with no due date (and no defer date still ahead);
        waiting is oldest wait first. A waiting task carries waiting_for:
        since (the ⏳ date), days, stale (past 7 days), on and on_path (who —
        the first person or company linked on the line).
      parameters: [{ name: view, in: query, schema: { type: string, enum: [inbox, today, upcoming, waiting, someday, logbook], default: today } }]
      responses: { "200": { description: "data: Task[]" } }
    post:
      summary: Quick capture — appends a task to today's daily note
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                due:
                  type: string
                  description: >
                    YYYY-MM-DD or a natural form (today, tomorrow, fri, +3d),
                    resolved server-side to ISO before it is written. An
                    unparseable date is a 400, never a guess and never
                    written to the file verbatim.
                defer:
                  type: string
                  description: hidden until this date; same formats as due
                waiting: { type: boolean, description: "mark waiting (⏳ stamped with today's date)" }
                waiting_on:
                  type: string
                  description: >
                    who it waits on, by name or alias; linked as [[Title]]
                    on the line unless already linked. Implies waiting.
      responses:
        "201": { description: "data: Task" }
        "400": { description: "VALIDATION_ERROR — missing text, or a date that could not be resolved" }
  /api/v1/tasks/{id}:
    patch:
      summary: Edit a task's dates/priority/waiting (surgical single-line rewrite)
      description: >
        Body fields: due, defer (YYYY-MM-DD; empty string clears; omitted
        leaves unchanged), priority (0 none, 1 high, 2 medium, 3 low),
        waiting (true stamps ⏳ with today's date; false removes marker and
        date), waiting_on (a name; links it and marks waiting), recur, text.
        The returned task's id may differ — ids are content-derived.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "data: Task" } }
  /api/v1/waiting:
    get:
      summary: Waiting view, grouped by who
      description: >
        Open waiting tasks grouped by who owes them — people and companies
        first, then other names, then unassigned (name ""). Oldest wait
        first within each group; stale counts tasks waiting more than 7 days.
      parameters: [{ name: area, in: query, schema: { type: string } }]
      responses: { "200": { description: "data: WaitingGroup[]" } }
  /api/v1/notes/from-task:
    post:
      summary: Turn a task into a note
      description: >
        Body {id, title?, area?}. Creates a note titled title (default: the
        task's first eight words) with the task's text as its body, filed in
        the source document's area (area when the source has none), and
        replaces the task's line with a plain bullet linking it. Needs the
        write scope — it creates a document.
      responses: { "201": { description: "data: Document (the new note)" }, "404": { description: no such task } }
  /api/v1/tasks/{id}/toggle:
    post:
      summary: Toggle a task's checkbox in its source markdown (surgical, one line)
      description: >-
        Optional body {"done": bool} sets that state instead of flipping; a task
        already in it is returned unchanged with no write.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "data: Task (note: the id may change — completion stamps alter the content hash)" } }
  /api/v1/daily/{date}:
    get:
      summary: The daily note for a date (404 if not started)
      parameters: [{ name: date, in: path, required: true, schema: { type: string, format: date } }]
      responses: { "200": { description: "data: Document" }, "404": { description: no note that day } }
    post:
      summary: Get-or-create the daily note
      responses: { "200": { description: "data: Document" } }
  /api/v1/today:
    get:
      summary: The composed home payload
      description: "date, daily (nullable), meetings, overdue, due_today, available, waiting, recent, events (today's calendar-feed occurrences)."
      responses: { "200": { description: "data: TodayPayload" } }
  /api/v1/calendar:
    get:
      summary: Month view of vault activity
      description: >
        One cell per day of the month: whether a daily note exists, documents
        touched (capped, newest first), meetings scheduled, and completed-task
        count. `month` is YYYY-MM and defaults to the current month.
      parameters: [{ name: month, in: query, schema: { type: string, example: "2026-09" } }]
      responses: { "200": { description: "data: CalendarMonth {month, prev, next, days[]}" } }
  /api/v1/calendar/feeds:
    get:
      summary: Subscribed ICS calendar feeds with their last fetch outcome
      description: >
        URLs are masked (scheme, host, last four characters): a feed URL is a
        bearer secret and is never returned once saved. error is the last
        failure, already free of the URL.
      responses: { "200": { description: "data: CalendarFeed[] — id, url (masked), last_attempt, last_success, error, failures, events" } }
    post:
      summary: Subscribe to an ICS feed and fetch it at once
      description: "Body {url}. https:// or webcal:// (rewritten to https). Stored 0600 in the state dir, apart from settings.json."
      responses:
        "201": { description: "data: CalendarFeed[] — the list, already showing whether the new feed fetched" }
        "400": { description: "VALIDATION_ERROR — not an http(s) URL, or already subscribed" }
  /api/v1/calendar/feeds/{id}:
    delete:
      summary: Unsubscribe; the feed's events vanish at once
      responses: { "204": { description: removed }, "404": { description: no such feed } }
  /api/v1/calendar/refresh:
    post:
      summary: Fetch every feed now, ignoring backoff
      responses: { "200": { description: "data: CalendarFeed[]" } }
  /api/v1/calendar/events:
    get:
      summary: Calendar-feed events over a range, in the configured time zone
      description: >
        Recurring series are expanded (RRULE, EXDATE, RECURRENCE-ID overrides).
        from and to are inclusive days — YYYY-MM-DD or natural forms (today,
        fri, +3d); from defaults to today, to to a week later, at most 92 days.
        Attendees are matched to person documents by email: (string or list)
        then by name; note_path is the meeting note made for the occurrence.
      parameters:
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses: { "200": { description: "data: CalendarEvent[]" } }
  /api/v1/calendar/events/note:
    post:
      summary: Make the meeting note for one event occurrence
      description: >
        Body {uid, date}. From the meeting template, dated at the event's start,
        matched attendees in people: as wikilinks, the rest on the Attendees
        line, event_uid in frontmatter. Idempotent per occurrence.
      responses:
        "201": { description: "data: Document — created" }
        "200": { description: "data: Document — the occurrence's existing note" }
        "404": { description: no such occurrence in the feeds }
  /api/v1/meeting-prep:
    get:
      summary: Per-attendee context for a meeting note or a calendar event
      description: >
        ?path= a meeting note, or ?event_uid= (with optional &date=) a calendar
        event. For each attendee with a person page: last meeting together,
        open tasks mentioning them, waiting (⏳) tasks mentioning them,
        company, recent non-meeting notes linking them; plus unmatched names.
      parameters:
        - { name: path, in: query, schema: { type: string } }
        - { name: event_uid, in: query, schema: { type: string } }
        - { name: date, in: query, schema: { type: string, format: date } }
      responses: { "200": { description: "data: MeetingPrep" }, "400": { description: neither path nor event_uid }, "404": { description: not found } }
  /api/v1/attachments:
    post:
      summary: Upload a file (multipart field `file`)
      description: Stored under vault attachments/YYYY/MM with a collision-resistant name; response includes ready-to-insert markdown.
      responses: { "201": { description: "data: {path, markdown}" } }
  /api/v1/files/{path}:
    get:
      summary: Serve a non-markdown vault file (images, PDFs)
      responses: { "200": { description: file bytes } }
  /api/v1/events:
    get:
      summary: SSE stream of index changes
      description: "Events named `doc` with JSON `{path, action: upsert|delete}` — invalidate caches on receipt."
      responses: { "200": { description: text/event-stream } }
  /api/v1/rename:
    post:
      summary: Move a document, optionally rewriting inbound links
      description: >
        Body {path, new_path, rewrite_links}. Only links that stop resolving
        (filename/path-shaped) are rewritten; title/alias links survive a
        move untouched. Returns {document, rewritten: [paths]}. 409 when
        new_path exists.
      responses: { "200": { description: "data: RenameResult" }, "409": { description: target exists } }
  /api/v1/agent-guidance:
    get:
      summary: The owner's guidance for AI agents
      description: >
        Stored as an ordinary vault document (AGENTS.md) and composed into the
        MCP server's `instructions`, so every connecting agent is told the
        house rules. Returns {path, text}.
      responses: { "200": { description: "data: AgentGuidanceResponse" } }
    put:
      summary: Replace the agent guidance (empty text removes it)
      responses: { "200": { description: "data: AgentGuidanceResponse" } }
  /api/v1/link:
    post:
      summary: Add or remove a wikilink in a list-valued frontmatter key
      description: >
        Body {path, key, target, remove?}. This is how entities relate:
        attendees on a meeting (`people`), a person's employer (`company`),
        projects, etc. Idempotent — linking the same target twice does not
        rewrite the file. Frontmatter wikilinks are indexed, so the link
        immediately appears as a backlink on the target.
      responses: { "200": { description: "data: Document" } }
  /api/v1/aliases:
    post:
      summary: Add a name a document answers to
      description: >
        Body {path, alias}. Appends to the document's `aliases:` frontmatter,
        keeping the aliases already there, so every [[alias]] link resolves to
        it. This is how an Unwritten name like "Frances" is settled against a
        likely match (people/frances-bagley.md) instead of becoming a
        duplicate page. Idempotent, case-insensitively.
      responses:
        "200": { description: "data: Document" }
        "400": { description: "VALIDATION_ERROR — empty, or contains [ ] | # ^" }
  /api/v1/unwritten:
    get:
      summary: Names the vault links to that have no document, most-referenced first
      description: >
        Each entry carries up to five `likely_matches` — existing people and
        companies the name probably means (first name, title prefix, or
        dotted initials like F.B.), strongest reason only.
      responses: { "200": { description: "data: Unwritten[] — name, refs, sources, likely_matches" } }
  /api/v1/decisions:
    get:
      summary: The decision log, newest first
      description: >
        Bullets under any heading named "Decisions" (kind inline — the meeting
        template has one) and documents tagged `decision` (kind record, text =
        title; the record's own singular "Decision" section is not counted
        again). Dated by the source's frontmatter `date:`, a daily note's day,
        else its mtime day. Each carries the people, companies and projects it
        is about: linked in the bullet or in the source's frontmatter.
      parameters:
        - { name: q, in: query, schema: { type: string }, description: "every word must appear in the decision or its source's title" }
        - { name: entity, in: query, schema: { type: string }, description: "vault path or name — decisions about it, or made in its page" }
        - { name: area, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        "200": { description: "data: Decision[] — kind, text, date, path, title, type, line, area, entities" }
        "400": { description: "VALIDATION_ERROR — entity names no document" }
  /api/v1/work-items:
    get:
      summary: Where work-item references link (Settings → Work items)
      description: >
        Returns {url_template}. When set, the app renders `AB#2433` and a purely
        numeric `#2433` as links to it with `{id}` replaced; "" leaves them as
        text. Numeric `#2433` is never a tag. Links only — nothing is fetched.
      responses: { "200": { description: "data: WorkItemSettings" } }
    put:
      summary: Set the work-item URL template ("" turns the links off)
      description: "Body {url_template} — an http(s) URL containing {id}."
      responses:
        "200": { description: "data: WorkItemSettings" }
        "400": { description: "VALIDATION_ERROR — no {id}, or not an http(s) URL" }
  /api/v1/shares:
    get:
      summary: List share links (revoked included — audit trail)
      responses: { "200": { description: "data: ShareInfo[] — token, doc_path, url, view_count, expires_at, revoked_at" } }
    post:
      summary: Create a share link for a document
      description: "Body {path, expires_in_days?}. The returned url is built from the configured base URL."
      responses: { "201": { description: "data: ShareInfo" } }
  /api/v1/shares/{token}:
    delete:
      summary: Revoke a share link
      responses: { "204": { description: revoked } }
  /api/v1/areas:
    get:
      summary: Areas documents are filed under (work, personal, …), seeded plus discovered, with counts
      description: >
        An area is the frontmatter `area:` key, lowercased. Daily notes carry
        none and belong to every area. Pass ?area=<name> to /documents,
        /tasks and /today to narrow them, or ?area=none for unclassified.
      responses: { "200": { description: "data: AreaCount[] — area, count, color, defined" } }
    put:
      summary: Replace the defined areas (Settings → Areas)
      description: >
        Body {areas: [{name, color}]}. Names normalize to lowercase and must be
        unique; color is one of slate, red, orange, amber, green, teal, blue,
        violet, pink. Renaming does not rewrite documents — frontmatter is the
        truth, and a document under the old name shows as a discovered area
        until re-filed.
      responses:
        "200": { description: "data: AreaCount[]" }
        "400": { description: "VALIDATION_ERROR — a duplicate, reserved, or empty name, or an unknown colour" }
  /api/v1/tags:
    get:
      summary: Every tag with its document count, most-used first
      responses: { "200": { description: "data: TagCount[] — tag, count" } }
  /api/v1/daily:
    get:
      summary: The journal's page of history — existing daily notes before a date, newest first
      parameters:
        - { name: before, in: query, schema: { type: string, format: date }, description: "defaults to today" }
        - { name: limit, in: query, schema: { type: integer, default: 30, maximum: 200 } }
      responses: { "200": { description: "data: Document[] (full documents, so they render without a second round trip)" } }
  /api/v1/audit:
    get:
      summary: What agents did — API-token and OAuth-client writes and MCP tool calls, newest first
      description: >
        The owner's own browser session is not audited (its autosaves would
        drown the log); everything an agent does is, with the outcome.
      parameters: [{ name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }]
      responses: { "200": { description: "data: AuditEntry[] — at, principal, action, path, detail, ok" } }
  /api/v1/tokens:
    get:
      summary: List API tokens (revoked included — audit trail)
      description: >
        Never returns the token itself: only its 8-char display prefix, since
        the plaintext is shown once at creation and only a hash is stored.
      responses: { "200": { description: "data: TokenInfo[] — id, name, prefix, scopes, created_at, expires_at, revoked_at, last_used_at" } }
    post:
      summary: Mint an API token
      description: >
        Body {name, scopes[], expires_in_days?}. Scopes must be a subset of
        read/write/tasks; an unknown scope is a 400 rather than a token that
        silently grants nothing. The response carries the plaintext exactly
        once and it is never retrievable again.
      responses:
        "201": { description: "data: NewToken — {token: TokenInfo, plaintext}" }
        "400": { description: "VALIDATION_ERROR — missing name, no scopes, or an unknown scope" }
  /api/v1/tokens/{prefix}:
    delete:
      summary: Revoke an API token by its display prefix (or numeric id)
      responses:
        "204": { description: revoked }
        "404": { description: no active token with that prefix }
  /api/v1/connected-apps:
    get:
      summary: List OAuth clients that have been granted access
      description: >
        Consented clients only — anyone can register via DCR, so unconsented
        registrations are not connections. active_grant reports whether the
        client still holds an unexpired, unrevoked refresh token.
      responses: { "200": { description: "data: ConnectedApp[] — client_id, name, scopes, consented_at, last_used_at, active_grant" } }
  /api/v1/connected-apps/{id}:
    delete:
      summary: Disconnect an app
      description: >
        Revokes every token the client holds, drops its outstanding
        authorization codes, and forgets the client, so reconnecting has to
        pass consent again.
      responses:
        "204": { description: disconnected }
        "404": { description: no connected app with that id }
  /api/v1/auth/status:
    get:
      summary: Passkey-mode status (unauthenticated)
      responses: { "200": { description: "data: {registered, authenticated}" } }
  /mcp:
    post:
      summary: MCP endpoint (Streamable HTTP)
      description: The agent interface — same service layer and permissions as this API.
      responses: { "200": { description: MCP protocol } }